Excels/requirements.txt
Django==1.9.2
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
django-celery==3.1.17
django-excel==0.0.4
et-xmlfile==1.0.1
jdcal==1.2
kombu==3.0.35
mysqlclient==1.3.7
odfpy==0.9.6
openpyxl==2.3.5
pyexcel==0.2.4
pyexcel-io==0.2.1
pyexcel-ods==0.2.0
pyexcel-webio==0.0.7
pyexcel-xls==0.2.1
pyexcel-xlsx==0.2.1
pytz==2016.6.1
texttable==0.8.4
xlrd==1.0.0
